+++ R/restrict_environment.R
@@ -1,16 +1,16 @@
 
-#' Build a new funcion with a smaller environment
+#' Build a new function with a smaller environment
 #'
 #' As standard feature, R include in the environment of a function
 #' all the variables that are available when the function is created.
-#' This, however is prompt to leak refenrencewhen you have a factory
+#' This, however is prompt to leak reference when you have a factory
 #' of function and they are created within a list.. it will include all
 #' the component of the list in the function environment.
 #' To prevent that, the random generator functions are encapsulated with a
-#' restricted enviroment where only the variables that the function requires
+#' restricted environment where only the variables that the function requires
 #' to work are included
 #'
-#' @author  John Aponte
+#' @author  John J. Aponte
 #' @param f input function
 #' @param ... define the set of  variables to be included as variable = value.
 #' @return new function with a restricted environment