+++ tests/testthat/test_methods.R
@@ -65,3 +65,75 @@ test_that("The dimnames works well" , {
   myDistr <- new_NORMAL(0,1)
   expect_equal(dimnames(myDistr), "rvar")
 })
+
+
+test_that("The Summary does not change the system seed", {
+  d1 <- new_NORMAL(0,1)
+  # Setup the fixture
+  if (exists(".Random.seed", .GlobalEnv))
+    oldseed <- .GlobalEnv$.Random.seed
+  else
+    oldseed <- NULL
+  #test start
+  set.seed(123456789)
+  v1 <- runif(1000)
+  v2 <- runif(1000)
+  set.seed(123456789)
+  v3 <- runif(1000)
+  summary(d1)
+  v4 <- runif(1000)
+  set.seed(123456789)
+  v5 <- runif(1000)
+  rfunc(d1,10000)
+  v6 <- runif(1000)
+  expect_equal(v1,v3)
+  expect_equal(v1,v5)
+  expect_equal(v2,v4)
+  expect_true(all(v2 == v4))
+  expect_false(all(v2 == v6))
+  # Return back the fixture
+  if (!is.null(oldseed))
+    .GlobalEnv$.Random.seed <- oldseed
+  else
+    rm(".Random.seed", envir = .GlobalEnv)
+})
+
+
+test_that("The seed only affects the latest created object", {
+  d1 <- new_NORMAL(0,1)
+  d2 <- new_NORMAL(0,1)
+  d3 <- new_NORMAL(0,1)
+  
+  # Setup the fixture
+  if (exists(".Random.seed", .GlobalEnv))
+    oldseed <- .GlobalEnv$.Random.seed
+  else
+    oldseed <- NULL
+  
+  set.seed(123456789)
+  d4 <- d1 + d2
+  set.seed(123456789)
+  d5 <- d1 + d3
+  
+  s1 <- summary(d4)
+  s2 <- summary(d5)
+  
+  expect_equal(s2,s2)
+  
+  # Return back the fixture
+  if (!is.null(oldseed))
+    .GlobalEnv$.Random.seed <- oldseed
+  else
+    rm(".Random.seed", envir = .GlobalEnv)
+  
+})
+
+
+test_that("The set_seed function", {
+  d1 <- new_NORMAL(1,0)
+  s1 <- d1$seed
+  d2 <- set_seed(d1,123456790)
+  s2 <- d2$seed
+  expect_equal(s2,123456790)
+  expect_false(s1 == s2)
+})